requirements.txt
lxml==6.0.3
python-docx==1.2.0
typing_extensions==4.15.0
